Trash Can, Item cube 모델 추가 + 타이틀 캔버스
merge conflict 해결 테스트 커밋입니다. + 타이틀 캔버스
