+++ Computer Virus Survivors/Assets/Scenes/MainScene/Back_Vid.renderTexture.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a55b8673b04304be39a5600e2702c067
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 8400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
