+++ Computer Virus Survivors/Assets/Scenes/MainScene/MainScene_Background.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 57db13beabb984cada6f60f069ea464d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
