+++ Computer Virus Survivors/Assets/Scenes/MainScene/UIFollowMouse.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a9d9a3b68278b4e70a116af45d0d897e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
