+++ Computer Virus Survivors/Assets/Scenes/MainScene/back_video.mp4.meta
@@ -1,0 +1,18 @@
+fileFormatVersion: 2
+guid: 6395b1f14cce14bb69ba16bd25e2eecf
+VideoClipImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  frameRange: 0
+  startFrame: -1
+  endFrame: -1
+  colorSpace: 0
+  deinterlace: 0
+  encodeAlpha: 0
+  flipVertical: 0
+  flipHorizontal: 0
+  importAudio: 1
+  targetSettings: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
